% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bibtex_citations.R
\name{check_bibtex_citations}
\alias{check_bibtex_citations}
\title{Check BibTeX Citations}
\usage{
check_bibtex_citations(citation, package)
}
\arguments{
\item{citation}{A BibTeX citation of an R package}

\item{package}{An R package to be compared to the citation}
}
\value{
Any issues with the citation given when compared to the package's citation
}
\description{
Check BibTeX Citations
}
\examples{

check_bibtex_citations(citation, "tidyverse")

check_bibtex_citation(citation, "dplyr")



}
