% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bibtex_file.R
\name{check_bibtex_file}
\alias{check_bibtex_file}
\title{Check Bibtex File}
\usage{
check_bibtex_file(file, citations)
}
\arguments{
\item{file}{File containing BibTeX citations.}

\item{citations}{A vector of packages, in the order they appear in your citation file.}
}
\value{
Any issues with the citation given when compared to the package's citation.
}
\description{
This function goes through a file containing BibTeX citations and compares the citations to
a vector of citations given.
}
\details{
NOTE: The order of the citations vector must be the same as the order
of the citations in the file. For example, if the citations vector contains 2 packages and is
ordered as c(PackageA, PackageB), PackageA must be the first citation to appear in the citation
file.
}
\examples{

check_bibtex_file("references.bib", c("tidyverse", "Lahman"))


check_bibtex_file("references.bib", "knitr")

}
