% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_citation_file.R
\name{check_citation_file}
\alias{check_citation_file}
\title{Check Citation File}
\usage{
check_citation_file(file, packages)
}
\arguments{
\item{file}{A Bibtex file containing the package citations.}

\item{packages}{A list of packages used in the project.}
}
\value{
Whether or not the listed packages are found in the given file.
}
\description{
Check Citation File
}
\examples{
check_citation_file("references.bib", c("tidyverse", "Lahman"))

check_citation_file("ref.bib", "ggplot2")

}
