% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_citations.R
\name{check_citations}
\alias{check_citations}
\title{Check Citations}
\usage{
check_citations(citation, package, format = "bibtex")
}
\arguments{
\item{citation}{A citation for an R package.}

\item{package}{An R package name.}
}
\value{
TRUE if entries are the same, FALSE if there are differences
}
\description{
Check Citations
}
\examples{
check_citations(citation, "Lahman", format = "text")


}
