% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_citations.R
\name{create_citations}
\alias{create_citations}
\title{create_citations}
\usage{
create_citations(packages, format = "bibtex", include_r = TRUE)
}
\arguments{
\item{packages}{A list of packages that you have used or plan to use in your document.}

\item{format}{The type of citations you want to return. Returns bibtex by default, use `format = text` to return text citations.}

\item{include_r}{If set to TRUE, will create a citation for R.}
}
\value{
A list of citations for the packages included in a document.
}
\description{
This function creates citations for multiple R packages at once. By default, this function will create
a citation for R, along with the other functions you include. When creating the citations, you
can choose to either receive text citations or have them formatted in BibTeX. The function will return
citations in the console pane in your RStudio window.
}
\examples{
create_citations(packages = list("tidyverse", "DoSStoolkit"),
 format = "bibtex")

create_citations(list("Lahman", "ggplot2", "forcats"),
"text")

create_citations(list("tidyverse"))

}
