% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bibtex_file.R
\name{find_refs}
\alias{find_refs}
\title{Check Bibtex File}
\usage{
find_refs(file)
}
\arguments{
\item{file}{File containing BibTeX citations.}

\item{citations}{A vector of citations, in the order they appear in your citation file.}
}
\value{
Any issues with the citation given when compared to the package's citation.
}
\description{
Check Bibtex File
}
\examples{

check_bibtex_file("references.bib", c("tidyverse", "Lahman"))


check_bibtex_file("references.bib", "knitr")

}
