% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_citation_file.R
\name{check_citation_file}
\alias{check_citation_file}
\title{Check Citation File}
\usage{
check_citation_file(file, packages)
}
\arguments{
\item{file}{A Bibtex file containing the package citations.}

\item{packages}{A list of packages used in the project.}
}
\value{
Whether or not the listed packages are found in the given file.
}
\description{
This function looks at your BibTeX file that contains citations. It takes in a file name
and a list of packages. The function will look through the citation file and check if the
packages listed are found in the citation file. Packages that are both listed in the function
and found in the citation file will be printed in the console.
}
\examples{
check_citation_file("references.bib", c("tidyverse", "Lahman"))

check_citation_file("ref.bib", "ggplot2")

}
