% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_citations.R
\name{check_citations}
\alias{check_citations}
\title{Check Citations}
\usage{
check_citations(citation, package, format = "bibtex")
}
\arguments{
\item{citation}{A citation for an R package.}

\item{package}{An R package name.}
}
\value{
TRUE if entries are the same, FALSE if there are differences
}
\description{
Checks if the text citation given matches the citation given by the citation function in R.
If the citations are different, the function will return FALSE, if they are the same it will
return TRUE.
}
\examples{
check_citations(citation, "tidyverse", format = "text")


}
