% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_refs.R
\name{find_refs}
\alias{find_refs}
\title{Find Refs}
\usage{
find_refs(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
This is a helper function for the `check_bibtex_file` function.
}
\examples{


}
